[Misc] Move dataset_tests to an independent directory. (#6617)
