+++ tests/python/common/data/test_data.py
@@ -1759,29 +1759,6 @@ def test_as_nodepred2():
 
 
 @unittest.skipIf(
-    dgl.backend.backend_name != "pytorch", reason="ogb only supports pytorch"
-)
-@unittest.skipIf(dgl.backend.backend_name == "mxnet", reason="Skip MXNet")
-def test_as_nodepred_ogb():
-    from ogb.nodeproppred import DglNodePropPredDataset
-
-    ds = data.AsNodePredDataset(
-        DglNodePropPredDataset("ogbn-arxiv"), split_ratio=None, verbose=True
-    )
-    split = DglNodePropPredDataset("ogbn-arxiv").get_idx_split()
-    train_idx, val_idx, test_idx = split["train"], split["valid"], split["test"]
-    assert F.array_equal(ds.train_idx, F.tensor(train_idx))
-    assert F.array_equal(ds.val_idx, F.tensor(val_idx))
-    assert F.array_equal(ds.test_idx, F.tensor(test_idx))
-    # force generate new split
-    ds = data.AsNodePredDataset(
-        DglNodePropPredDataset("ogbn-arxiv"),
-        split_ratio=[0.7, 0.2, 0.1],
-        verbose=True,
-    )
-
-
-@unittest.skipIf(
     F._default_context_str == "gpu",
     reason="Datasets don't need to be tested on GPU.",
 )
@@ -1808,26 +1785,6 @@ def test_as_linkpred():
     assert ds.test_edges[0][0].shape[0] == 2112
     # negative samples, not guaranteed to be ratio 2, so the assert is in a relaxed range
     assert 4000 < ds.test_edges[1][0].shape[0] <= 4224
-
-
-@unittest.skipIf(
-    dgl.backend.backend_name != "pytorch", reason="ogb only supports pytorch"
-)
-def test_as_linkpred_ogb():
-    from ogb.linkproppred import DglLinkPropPredDataset
-
-    ds = data.AsLinkPredDataset(
-        DglLinkPropPredDataset("ogbl-collab"), split_ratio=None, verbose=True
-    )
-    # original dataset has 46329 test edges
-    assert ds.test_edges[0][0].shape[0] == 46329
-    # force generate new split
-    ds = data.AsLinkPredDataset(
-        DglLinkPropPredDataset("ogbl-collab"),
-        split_ratio=[0.7, 0.2, 0.1],
-        verbose=True,
-    )
-    assert ds.test_edges[0][0].shape[0] == 235812
 
 
 @unittest.skipIf(
@@ -1989,25 +1946,6 @@ def test_as_graphpred_reprocess():
     # invalid cache, re-read
     ds = data.AsGraphPredDataset(data.BA2MotifDataset(), [0.1, 0.1, 0.8])
     assert len(ds.train_idx) == int(len(ds) * 0.1)
-
-
-@unittest.skipIf(
-    dgl.backend.backend_name != "pytorch", reason="ogb only supports pytorch"
-)
-def test_as_graphpred_ogb():
-    from ogb.graphproppred import DglGraphPropPredDataset
-
-    ds = data.AsGraphPredDataset(
-        DglGraphPropPredDataset("ogbg-molhiv"), split_ratio=None, verbose=True
-    )
-    assert len(ds.train_idx) == 32901
-    # force generate new split
-    ds = data.AsGraphPredDataset(
-        DglGraphPropPredDataset("ogbg-molhiv"),
-        split_ratio=[0.6, 0.2, 0.2],
-        verbose=True,
-    )
-    assert len(ds.train_idx) == 24676
 
 
 if __name__ == "__main__":
